Create linked risks table body and clear unused code
