+++ Clients/src/presentation/components/LinkedRisks/index.tsx
@@ -13,6 +13,8 @@ import { ReactComponent as CheckboxFilled } from "../../assets/icons/checkbox-fi
 import { ReactComponent as SelectorVertical } from '../../assets/icons/selector-vertical.svg'
 import { ProjectRisk } from '../../../domain/types/ProjectRisk';
 
+import LinkedRisksTable from '../Table/LinkedRisksTable';
+
 import {
   textfieldStyle,
   tableWrapper,
@@ -86,57 +88,7 @@ const LinkedRisksPopup: React.FC<LinkedRisksModalProps> = ({
             />
           </Stack>
         </Stack>
-        <Stack>
-          <TableContainer>
-            <Table
-              sx={{
-                ...singleTheme.tableStyles.primary.frame,
-                ...tableWrapper
-              }}
-            >
-              <TableHeader columns={TITLE_OF_COLUMNS} />
-              {projectRisks.length > 0 ? 
-                <>
-                  {filteredRisks.length > 0 ? 
-                    <RiskTableBody 
-                      rows={filteredRisks} 
-                      setCurrentPagingation={setCurrentPagingation}
-                      page={currentPage}
-                    />
-                  : <>
-                    <TableRow>
-                      <TableCell
-                        colSpan={TITLE_OF_COLUMNS.length}
-                        align="center"
-                        sx={emptyData}
-                      >
-                        <img src={placeholderImage} alt="Placeholder" />
-                        <Typography sx={styles.textBase}>
-                          No risks found in database
-                        </Typography>
-                      </TableCell>
-                    </TableRow>
-                  </>}
-                </> 
-                : <>
-                  <TableBody>
-                    <TableRow>
-                      <TableCell
-                        colSpan={TITLE_OF_COLUMNS.length}
-                        align="center"
-                        sx={emptyData}
-                      >
-                        <img src={placeholderImage} alt="Placeholder" />
-                        <Typography sx={styles.textBase}>
-                          There is currently no risk in this project.
-                        </Typography>
-                      </TableCell>
-                    </TableRow>
-                  </TableBody>
-                </>}
-            </Table>
-          </TableContainer>
-        </Stack>
+        <LinkedRisksTable projectRisksGroup={projectRisks} filteredRisksGroup={filteredRisks}/>
       </Stack>
       <Stack sx={{ display: 'flex', flexDirection: 'row', justifyContent: 'flex-end'}}>
         <Button 
@@ -151,127 +103,6 @@ const LinkedRisksPopup: React.FC<LinkedRisksModalProps> = ({
         />
       </Stack>
     </Stack>
-  )
-}
-
-const RiskTableBody: React.FC<TableProps> = ({
-  rows,
-  page,
-  setCurrentPagingation
-}) => {
-  const cellStyle = singleTheme.tableStyles.primary.body.cell;
-  const [rowsPerPage, setRowsPerPage] = useState(5);
-  const [checkedRows, setCheckedRows] = useState<number[]>([]);
-  const theme = useTheme();  
-
-  const handleChangePage = useCallback((_: unknown, newPage: number) => {
-    setCurrentPagingation(newPage);
-  }, [setCurrentPagingation]);
-
-  const handleChangeRowsPerPage = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      setRowsPerPage(parseInt(event.target.value, 10));
-      setCurrentPagingation(0);
-    },
-    [setRowsPerPage, setCurrentPagingation]
-  );
-
-  const handleRowClick = (riskId: number) => {
-    setCheckedRows((prev) =>
-      prev.includes(riskId)
-        ? prev.filter((i) => i !== riskId)
-        : [...prev, riskId]
-    );
-  };
-
-  return (
-    <>
-      <TableBody>
-        {rows &&
-          rows
-          .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-            .map((row: ProjectRisk, index: number) => (
-              <TableRow key={index} sx={singleTheme.tableStyles.primary.body.row} onClick={() => handleRowClick(row.id)}>
-                <TableCell sx={cellStyle}>
-                  <MuiCheckbox
-                    size="small"
-                    id="auto-fill"
-                    checked={checkedRows.includes(row.id)}
-                    onChange={() => handleRowClick(row.id)}
-                    onClick={(e) => e.stopPropagation()}  
-                    checkedIcon={<CheckboxFilled />}
-                    icon={<CheckboxOutline />}
-                    sx={{
-                      borderRadius: "4px",
-                      "&:hover": { backgroundColor: "transparent" },
-                      "& svg": { width: "small", height: "small" },
-                      "& .MuiTouchRipple-root": {
-                        display: "none",
-                      },
-                    }}
-                  />
-                </TableCell>
-                <TableCell sx={cellStyle}>
-                  {row.id ? row.id : page * rowsPerPage + index + 1}
-                </TableCell>
-                <TableCell>
-                  {row.risk_name ? row.risk_name : '-'}
-                </TableCell>
-                <TableCell sx={{maxWidth: '300px'}}>
-                  {row.risk_description ? row.risk_description : '-'}
-                </TableCell>
-                <TableCell sx={cellStyle}>
-                  {row.risk_severity ? <RiskChip label={row.risk_severity} /> : '-'}
-                </TableCell>
-                <TableCell sx={cellStyle}>
-                  {row.likelihood ? row.likelihood : '-'}
-                </TableCell>
-                <TableCell sx={cellStyle}>
-                  {row.risk_category ? row.risk_category : '-'}
-                </TableCell>
-              </TableRow>
-            ))}
-      </TableBody>
-      <TableFooter>
-        <TableRow sx={{
-          '& .MuiTableCell-root.MuiTableCell-footer': {
-            paddingX: theme.spacing(8),
-            paddingY: theme.spacing(4),
-          }}}>
-          <TablePagination
-            count={rows?.length}
-            page={page}
-            onPageChange={handleChangePage}
-            rowsPerPage={rowsPerPage}
-            rowsPerPageOptions={[5, 10, 15, 20, 25]}
-            onRowsPerPageChange={handleChangeRowsPerPage}
-            ActionsComponent={(props) => <TablePaginationActions {...props} />}
-            labelRowsPerPage="Risks per page"
-            labelDisplayedRows={({ page, count }) =>
-              `Page ${page + 1} of ${Math.max(0, Math.ceil(count / rowsPerPage))}`
-            }
-            sx={paginationStyle}
-            slotProps={{
-              select: {
-                MenuProps: {
-                  keepMounted: true,
-                  PaperProps: {
-                    className: "pagination-dropdown",
-                    sx: paginationDropdown,
-                  },
-                  transformOrigin: { vertical: "bottom", horizontal: "left" },
-                  anchorOrigin: { vertical: "top", horizontal: "left" },
-                  sx: { mt: theme.spacing(-2) },
-                },
-                inputProps: { id: "pagination-dropdown" },
-                IconComponent: SelectorVertical,
-                sx: paginationSelect,
-              },
-            }}
-          />
-        </TableRow>
-      </TableFooter>
-    </>
   )
 }
